+++ context.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2017 ArangoDB GmbH, Cologne, Germany
+// Copyright 2017-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,9 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Ewout Prangsma
-// Author Tomasz Mielech <tomasz@arangodb.com>
 //
 
 package driver
@@ -70,6 +67,8 @@ const (
 	keyDropCollections          ContextKey = "arangodb-drop-collections"
 	keyDriverFlags              ContextKey = "arangodb-driver-flags"
 	keyRefillIndexCaches        ContextKey = "arangodb-driver-refill-index-caches"
+	keyAsyncRequest             ContextKey = "arangodb-async-request"
+	keyAsyncID                  ContextKey = "arangodb-async-id"
 )
 
 type OverwriteMode string
@@ -298,6 +297,16 @@ func WithRefillIndexCaches(parent context.Context, value bool) context.Context {
 	return context.WithValue(contextOrBackground(parent), keyRefillIndexCaches, value)
 }
 
+// WithAsync is used to configure a context to make an async operation - requires Connection with Async wrapper!
+func WithAsync(parent context.Context) context.Context {
+	return context.WithValue(contextOrBackground(parent), keyAsyncRequest, true)
+}
+
+// WithAsyncID is used to check an async operation result - requires Connection with Async wrapper!
+func WithAsyncID(parent context.Context, asyncID string) context.Context {
+	return context.WithValue(contextOrBackground(parent), keyAsyncID, asyncID)
+}
+
 type contextSettings struct {
 	Silent                   bool
 	WaitForSync              bool
@@ -368,6 +377,7 @@ func applyContextSettings(ctx context.Context, req Request) contextSettings {
 	if ctx == nil {
 		return result
 	}
+
 	// Details
 	if v := ctx.Value(keyDetails); v != nil {
 		if details, ok := v.(bool); ok {
@@ -550,6 +560,13 @@ func applyContextSettings(ctx context.Context, req Request) contextSettings {
 		}
 	}
 
+	// AsyncID
+	if v := ctx.Value(keyAsyncID); v != nil {
+		if asyncID, ok := v.(string); ok {
+			req.SetHeader("x-arango-async-id", asyncID)
+		}
+	}
+
 	return result
 }
 
@@ -593,4 +610,34 @@ func withDocumentAt(ctx context.Context, index int) (context.Context, error) {
 	}
 
 	return ctx, nil
+}
+
+//
+// READ METHODS
+//
+
+// IsAsyncRequest returns true if the given context is an async request.
+func IsAsyncRequest(ctx context.Context) bool {
+	if ctx != nil {
+		if v := ctx.Value(keyAsyncRequest); v != nil {
+			if isAsync, ok := v.(bool); ok && isAsync {
+				return true
+			}
+		}
+	}
+
+	return false
+}
+
+// HasAsyncID returns the async Job ID from the given context.
+func HasAsyncID(ctx context.Context) (string, bool) {
+	if ctx != nil {
+		if q := ctx.Value(keyAsyncID); q != nil {
+			if v, ok := q.(string); ok {
+				return v, true
+			}
+		}
+	}
+
+	return "", false
 }