+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add support for getting license
 - Add support for Raw Authentication in VST (support external jwt token as raw element)
 - Fix race when using WithRawResponse/WithResponse context with agencyConnection 
+- [V1] Async Client
 
 ## [1.6.0](https://github.com/arangodb/go-driver/tree/v1.6.0) (2023-05-30)
 - Add ErrArangoDatabaseNotFound and IsExternalStorageError helper to v2